Growing crops and harvesting
-Crops now grow on a timer when planted and watered
-Crops now a seperate class, plots "have" crops
-Can only control camera when holding the "c" key
-Can now interact with mouse button 1
-Can now scroll thorugh items on hotbar with scroll wheel

+++ Sandbox 1/Assets/C# Scripts/Grid.cs
@@ -13,6 +13,7 @@ public class Grid : MonoBehaviour
     Vector3 plotSize;
     private Dictionary<string, GameObject> plotDictionary;
     public bool DrawGizmos = false;
+    public bool plotDebugger = false;
 
 
 
@@ -24,12 +25,14 @@ public class Grid : MonoBehaviour
 
     private void Update()
     {
-        if (Input.GetKeyDown("3"))
+        if (plotDebugger)
         {
-            destroyPlots();
-            DrawPlots();
+            if (Input.GetKeyDown("3"))
+            {
+                destroyPlots();
+                DrawPlots();
+            }
         }
- 
     }
 
     //Draw plots in the grid
